float struggle begins

+++ libft/ft_ftoa.c
@@ -6,20 +6,20 @@
 /*   By: mviinika <mviinika@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/15 20:54:44 by mviinika          #+#    #+#             */
-/*   Updated: 2022/03/16 19:40:00 by mviinika         ###   ########.fr       */
+/*   Updated: 2022/03/22 21:55:21 by mviinika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_ftoa(float num, int afterpoint)
+char	*ft_ftoa(long double num, int afterpoint)
 {
-	int		l_dot;
-	float	r_dot;
-	char	*res;
+	int			l_dot;
+	long double	r_dot;
+	char		*res;
 
 	l_dot = (int)num;
-	r_dot = num - (float)l_dot;
+	r_dot = num - (long double)l_dot;
 	res = ft_strnew(20);
 	res = ft_itoa(l_dot);
 	if (afterpoint != 0)
